+++ flash/image/segmentation/input.py
@@ -20,10 +20,9 @@ from pytorch_lightning.utilities import rank_zero_warn
 from flash.core.data.io.input import DataKeys, ImageLabelsMap, Input
 from flash.core.data.utilities.paths import filter_valid_files, PATH_TYPE
 from flash.core.data.utilities.samples import to_samples
-from flash.core.data.utils import image_default_loader
 from flash.core.integrations.fiftyone.utils import FiftyOneLabelUtilities
 from flash.core.utilities.imports import _FIFTYONE_AVAILABLE, _TORCHVISION_AVAILABLE, lazy_import
-from flash.image.data import ImageDeserializer, IMG_EXTENSIONS
+from flash.image.data import image_loader, ImageDeserializer, IMG_EXTENSIONS
 from flash.image.segmentation.output import SegmentationLabelsOutput
 
 SampleCollection = None
@@ -101,12 +100,12 @@ class SemanticSegmentationFilesInput(SemanticSegmentationInput):
         if mask_files is None:
             files = filter_valid_files(files, valid_extensions=IMG_EXTENSIONS)
         else:
-            files, masks = filter_valid_files(files, mask_files, valid_extensions=IMG_EXTENSIONS)
+            files, mask_files = filter_valid_files(files, mask_files, valid_extensions=IMG_EXTENSIONS)
         return to_samples(files, mask_files)
 
     def load_sample(self, sample: Dict[str, Any]) -> Dict[str, Any]:
         filepath = sample[DataKeys.INPUT]
-        sample[DataKeys.INPUT] = FT.to_tensor(image_default_loader(filepath))
+        sample[DataKeys.INPUT] = FT.to_tensor(image_loader(filepath))
         if DataKeys.TARGET in sample:
             sample[DataKeys.TARGET] = torchvision.io.read_image(sample[DataKeys.TARGET])[0]
         sample = super().load_sample(sample)
@@ -124,6 +123,7 @@ class SemanticSegmentationFolderInput(SemanticSegmentationFilesInput):
     ) -> List[Dict[str, Any]]:
         self.load_labels_map(num_classes, labels_map)
         files = os.listdir(folder)
+        files.sort()
         if mask_folder is not None:
             mask_files = os.listdir(mask_folder)
 
@@ -137,8 +137,10 @@ class SemanticSegmentationFolderInput(SemanticSegmentationFilesInput):
 
             files = [os.path.join(folder, file) for file in all_files]
             mask_files = [os.path.join(mask_folder, file) for file in all_files]
+            files.sort()
+            mask_files.sort()
             return super().load_data(files, mask_files)
-        return super().load_data(files)
+        return super().load_data([os.path.join(folder, file) for file in files])
 
 
 class SemanticSegmentationFiftyOneInput(SemanticSegmentationFilesInput):